Changes made: 1. In SAE J2735 Message Library (AsnJ2735Lib.cpp), removed encoding intersection MAP into multiple MAP messages with each message containing a subset of the intersection lane geometries that have the same speed limit. Persistence of speed limit follows the persistence of lane width. Same changes have applied to MAP decoder. 2. In SAE J2735 Message Library (AsnJ2735Lib.cpp), added decoding and encoding computed lanes. 3. In MAP Engine (locAware.cpp), changed LocAware::LocAware constructor to be consistent with changes made #1, i.e., removed the Boolean argument isSingleFrame from the constructor. 4. In MAP Engine (locAware.cpp), added construction of computed lane nodes from its reference lane, save computed lanes information into nmap file, and read nmap file with computed lanes information and encode it into MAP message.
Validation approach:
1. MAP Engine to read MAP payload file that was generated by the USDOT MAP Tool;
2. MAP Engine to decode the MAP payload and save the decoded results into a nmap file;
3. MAP Engine to read the nmap file and encode it into MAP payload; and
4. Compare the encoded MAP payload (step 4) with the original MAP payload (steep 1).

Validation results:
1. Added daisy-gavilan.map.payload that was generated by the USDOT MAP tool and contains computed lanes to tools/nmap directory;
2. Added nmap and MAP payload file generated by MAP Engine encoder to tools/output directory

+++ mrp/mapEngine/include/geoUtils.h
@@ -52,7 +52,7 @@ namespace GeoUtils
 
 	struct enuCoord_t
 	{
-		GeoUtils::point3D_t     pointECEF;        // ENU origin reference in ECEF
+		GeoUtils::point3D_t     pointECEF;        // ENU origin referenced in ECEF
 		GeoUtils::transMatrix_t transMatrix;
 	};
 
@@ -70,11 +70,11 @@ namespace GeoUtils
 		};
 		uint32_t length() const
 		{
-			return (static_cast<uint32_t>(round(std::sqrt(x  * x + y * y))));
+			return (static_cast<uint32_t>(round(std::sqrt((long)x * x + y * y))));
 		};
 		uint32_t distance2pt(const GeoUtils::point2D_t& p) const
 		{
-			return (static_cast<uint32_t>(round(std::sqrt((x - p.x) * (x - p.x) + (y - p.y) * (y - p.y)))));
+			return (static_cast<uint32_t>(round(std::sqrt((long)(x - p.x) * (x - p.x) + (y - p.y) * (y - p.y)))));
 		};
 		uint16_t direction2pt(const GeoUtils::point2D_t& p) const
 		{
@@ -108,9 +108,9 @@ namespace GeoUtils
 
 	struct projection_t
 	{
-		double t;       // along the line (unit-less)
-		double d;       // distance away from the line, positive means on the right-side of travel
-		double length;  // length of the line segment
+		double t;       // along the line segment (unit-less)
+		double d;       // distance away from the line segment in centimeters, positive means on the right-side of travel
+		double length;  // length of the line segment in centimeters
 		void reset(void)
 		{
 			t = 0.0;
